Graph w/ Adjacency Matrix
This is implementation of a graph can be weighted or not weighted, as well as directed or not directed.
There are a couple methods to take the class's adjacency list and return an adjacency matrix. For a non-weighted graph,
the adjacency matrix return is represented with the number one to represent an edge connecting two vertex. Starting at (0,0), "x" keys are in order top to bottom, and the "y" keys are order left to right. The weighted adjacency method is similar, but instead uses weights to signify an edge connecting two vertex rather than a "1". Both methods allow for named keys.