feat: add dijkstra using fib heap #11 (#39)
* feat: add fib heap skeleton #12

* feat: add fib heap functions find_min, insert, merge, extract_min #12

* feat: add decrease-key and delete #12

* style: run black on file

* doc: add fib heap to readme #12

* doc: update fib algorithm introduction #12

* fix: rework reference issue causing bug #25

* test(fib-heap): add test cases for fib-heap #13
added fouor complete test cases for the fib heap and one test case
that needs continued work. The test cases have a coverage of 63%
of the code of the fibonacci heap.

* test(fib-heap): fix bug in test case fib_heap resolve #27
resolve bug in test cases for ibonacci heap where a generator was
inserted into the fibonacci heap.

* fix(fib-heap): fix bug in merge fibonacci heap resolve #26
Added check for when the merged heap is empty, the function does
nothing. When the original heap is empty, the funcitno simply
replaces the original heap with the merged heap and switches the
min_nodes. When both heaps are non-empty, the function links the
root lists and updates the min_node.

* fix(fib-heap): refactor fib-heap merge #28
swapped places of the heap2_tail allocation in order to not
allocate a variable if it is not needed in the function execution

* fix(fib_heap): add update to total_nodes resolve #30
Add update to total nodes when an empty heap is merged with a non-
empty heap.

* test(fib-heap): add test for delete method #13
Added a test for the delete method which checks for two test cases
the first one being if one deletes a root element and making sure
the element is no longer in the heap. The other test case is for
deleting a child and making sure the node is no longer in the heap

* feat: add dijkstra's using fib heap #11

* style: remove unecessary comment in test

* test: add test cases for decrease_key method in Fibonacci heap #13 (#34)

* test: Add assertion to test decrease_key method #13

* Add check checks that the new heap is correct after decrease_key #13

* test: Add test case for checking Exception #13

* test: Add negative test for decrease_key #13

* test: Add test case for decreasing key of child #11

* test: Add test case for decreasing key of child #13

* test: Add test case with only one child #13

* doc: add docstring to fib heap dijkstra

* feat(fib-heap): add get_all_nodes function resolve #35 (#36)

* feat(fib-heap): add get_all_nodes function #35
Added function to recursively find all nodes of the fib-heap in
order to test that some specific nodes exist or that they do not

* doc: add doc strings to get all nodes #36

Co-authored-by: Johan von Hacht <JohanKJIP@gmail.com>

* style: remove unecessary code

* fix: correct fib heap iterate calls

* feat: implment dijkstra using decrease_key #11

* fix: fib heap now returns node on insert

* style: remove redundant code

* style: use correct iterate call

Co-authored-by: jonasjungaker <32596748+jonasjungaker@users.noreply.github.com>
Co-authored-by: Emma OIsson <51324301+emmariaolsson@users.noreply.github.com>