fix_rotate.py (#778)
* fix_rotate.py

The way it was, you couldn't pass in a rotation higher than the length of the string doubled without getting a blank string in return. This way allows you to pass in any positive integer and get a result.

Also,
The last line of the description said that it took 2 strings and returned a boolean. This is not correct. The first line of the comment is correct.

* Update rotate.py

After submitting my last change, I realized it would have been better to not keep concatenating the string.

* Updated Comments rotate.py